merged-project-flask/requirements.txt
absl-py==2.3.1
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
datasets==2.21.0
dill==0.3.7
distro==1.9.0
facenet-pytorch==2.5.3
filelock==3.18.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.3
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.73.1
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
keras==3.10.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.4.8
libclang==18.1.1
lxml==6.0.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.26.1
matplotlib==3.10.3
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
modelscope==1.28.0
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.15
namex==0.1.0
networkx==3.5
numpy==1.26.4
openai==1.97.0
opencv-contrib-python==4.11.0.86
opencv-python==4.5.5.64
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
propcache==0.3.2
protobuf==4.25.8
psutil==6.1.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.19.2
PyMySQL==1.0.2
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentencepiece==0.2.0
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
SpeechRecognition==3.8.1
SQLAlchemy==1.4.46
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
Werkzeug==2.0.1
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
